Patch 1.2
Massive update to the code. Completely rearranged the drivetrain
according to the way it was written in 2014. I cannot see straight right
now. Added buttons.h, and oimap.h. This is what Jeff was telling us
about last year, it is more organized than shoving everything in the
oi.h and oi.cpp. Instead of shoving everything in the OI, the OIMap has
all of the buttons defined relating back to the xb360 map. The OIMap
makes the XB360 more usable and cleaner. The Axis Scaling Factor is also
declared in the OIMap. Buttons.h has which pad is being used, the "when
pressed" or "while held," and the Command. This is much cleaner. 

Code has not been tested on a machine yet but I think we need to start
by building it on a laptop. After building eclipse spits out a cRIO
debug file beside the SRC file. Whoever gets a chance to push first, we
need to try deploying that file specifically instead of the entire
project.

LIT