Added drive encoders.
Set distance per pulse = 8*pi. Not sure of the exact size of the wheels
my guess is eight inches. Will confirm next meeting. Not sure of how to
incorporate the turret motors or the gatling. Gatling will most likely
be controlled by a snow blower motor. Turret will most likely be
controlled by a globe motor then a seperate motor for the up and down
action of the cannon. We are going to begin the tee shirt launcher soon,
feel free to commit to this repo whenever possible.